many to many added, split employer info. gotta create daos and controllers and html
