+++ src/main/java/org/stlyouthjobs/models/NewEmployer.java
@@ -16,7 +16,7 @@ import static javax.persistence.CascadeType.PERSIST;
 import static javax.persistence.FetchType.EAGER;
 
 @Entity
-@Table(name = "USER")
+@Table(name = "NEW_EMPLOYER")
 public class NewEmployer {
 
     @NotNull
@@ -41,76 +41,30 @@ public class NewEmployer {
             message = "Invalid phone number")
     private Long phoneNumber;
 
-    @NotNull
-    @Email
-    private String email;
-
-    @NotNull
-    @Email
-    private String confirmEmail;
-
-    @NotNull
-    @Size(max = 2500, message = "This field is required.")
-    private String password;
-
-    @NotNull
-    @Size(max = 2500, message = "This field is required.")
-    private String confirmPassword;
-
-    @NotNull
-    @Size(min=1, max=10, message = "Street number must have at least one number.")
-    private Integer streetNumber;
-
-    @NotNull
-    @Size(max = 2500, message = "This field is required.")
-    private String streetName;
-
-    @NotNull
-    @Size(max = 2500, message = "This field is required.")
-    private String city;
-
-    @NotNull
-    @Size(max = 9, message = "This field is required.")
-    private Integer zipCode;
-
-    @NotNull
-    @Size(max = 2500, message = "This field is required.")
-    private String neighborhood;
 
     @Id
     @GeneratedValue
-    private String uniqueID = UUID.randomUUID().toString();
+    private int id;
 
     @ManyToMany(mappedBy =  "newEmployers", cascade = PERSIST, fetch = EAGER)
-    private Set<User> users = new HashSet<>();
+    private Set<EmployerUser> employerUsers = new HashSet<>();
 
-    public Set<User> getUsers() {
-        return users;
+    public Set<EmployerUser> getEmployerUsers() {
+        return employerUsers;
     }
 
-    public void addUser(User user) {
-        users.add(user);
-        user.getNewEmployer().add(this);
+    public void addEmployerUser(EmployerUser employerUser) {
+        employerUsers.add(employerUser);
+        employerUser.getNewEmployer().add(this);
     }
     public NewEmployer () {}
 
-    public NewEmployer(String organizationName, String contactName, String website, Long ein, Long phoneNumber,
-                       String email, String confirmEmail, String password, String confirmPassword, Integer streetNumber,
-                       String streetName, String city, Integer zipCode, String neighborhood) {
+    public NewEmployer(String organizationName, String contactName, String website, Long ein, Long phoneNumber) {
         this.organizationName = organizationName;
         this.contactName = contactName;
         this.website = website;
         this.ein = ein;
         this.phoneNumber = phoneNumber;
-        this.email = email;
-        this.confirmEmail = confirmEmail;
-        this.password = password;
-        this.confirmPassword= confirmPassword;
-        this.streetNumber = streetNumber;
-        this.streetName = streetName;
-        this.city = city;
-        this.zipCode = zipCode;
-        this.neighborhood = neighborhood;
     }
     public int getId() {
         return id;
@@ -156,75 +110,4 @@ public class NewEmployer {
         this.phoneNumber = phoneNumber;
     }
 
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getConfirmEmail() {
-        return confirmEmail;
-    }
-
-    public void setConfirmEmail(String confirmEmail) {
-        this.confirmEmail = confirmEmail;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getConfirmPassword() {
-        return confirmPassword;
-    }
-
-    public void setConfirmPassword(String confirmPassword) {
-        this.confirmPassword = confirmPassword;
-    }
-
-    public Integer getStreetNumber() {
-        return streetNumber;
-    }
-
-    public void setStreetNumber(Integer streetNumber) {
-        this.streetNumber = streetNumber;
-    }
-
-    public String getStreetName() {
-        return streetName;
-    }
-
-    public void setStreetName(String streetName) {
-        this.streetName = streetName;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public Integer getZipCode() {
-        return zipCode;
-    }
-
-    public void setZipCode(Integer zipCode) {
-        this.zipCode = zipCode;
-    }
-
-    public String getNeighborhood() {
-        return neighborhood;
-    }
-
-    public void setNeighborhood(String neighborhood) {
-        this.neighborhood = neighborhood;
-    }
 }
